+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.1.0 (in development)
 ------------------------------------------------------------------------
+- Feature: Allow player to remove roads that are in use.
 - Feature: Towns can now always be renamed (As seen in OpenTTD).
 - Feature: Vehicle breakdowns can now be disabled (As seen in OpenTTD).
 - Feature: Playable in a resizable window.